Move parameterize_exchanges method into model
Clean up imports and method from table
