+++ activity_browser/ui/tables/parameters.py
@@ -1,12 +1,9 @@
 # -*- coding: utf-8 -*-
 from asteval import Interpreter
-import brightway2 as bw
-from bw2data.parameters import ActivityParameter
 from PySide2.QtCore import Slot
 from PySide2.QtGui import QContextMenuEvent, QDragMoveEvent, QDropEvent
 from PySide2.QtWidgets import QAction, QMenu, QMessageBox
 
-from ...bwutils import commontasks as bc
 from ...settings import project_settings
 from ...signals import signals
 from ..icons import qicons
@@ -256,25 +253,3 @@ class ExchangesTable(ABDictTreeView):
         self.model = ParameterTreeModel(parent=self)
         self.setModel(self.model)
         self.model.updated.connect(self.custom_view_sizing)
-
-    def _connect_signals(self):
-        super()._connect_signals()
-        signals.exchange_formula_changed.connect(self.parameterize_exchanges)
-
-    @Slot(tuple, name="parameterizeExchangesForKey")
-    def parameterize_exchanges(self, key: tuple) -> None:
-        """ Used whenever a formula is set on an exchange in an activity.
-
-        If no `ActivityParameter` exists for the key, generate one immediately
-        """
-        group = bc.build_activity_group_name(key)
-        if not (ActivityParameter.select()
-                .where(ActivityParameter.group == group).count()):
-            signals.add_activity_parameter.emit(key)
-
-        act = bw.get_activity(key)
-        with bw.parameters.db.atomic():
-            bw.parameters.remove_exchanges_from_group(group, act)
-            bw.parameters.add_exchanges_to_group(group, act)
-            ActivityParameter.recalculate_exchanges(group)
-        signals.parameters_changed.emit()
